requirements.txt
streamlit
yfinance
plotly
google-genai
pandas
python-docx